refactor: separate state persistence and history
